+++ ams/opt/__init__.py
@@ -8,5 +8,5 @@ from ams.opt.exprcalc import ExpressionCalc  # NOQA
 from ams.opt.param import Param  # NOQA
 from ams.opt.constraint import Constraint  # NOQA
 from ams.opt.objective import Objective  # NOQA
-from ams.opt.expr import Expression  # NOQA
+from ams.opt.expression import Expression  # NOQA
 from ams.opt.omodel import OModel  # NOQA
